+++ dd-qa-cinema-react-app/package.json
@@ -19,7 +19,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "connect": "node ./src/API/Run.js"
+    "connect": "node ./src/API/APICall.js"
   },
   "eslintConfig": {
     "extends": [
