+++ dd-qa-cinema-react-app/src/API/APICall.js
@@ -2,9 +2,9 @@ const connect = require("express");
 const bodyParser = require('body-parser');
 const app = connect();
 
-const FilmsAPI = require("./Films.js");
-const BookingsAPI = require("./Bookings.js");
-const ScreeningAPI = require("./Screenings.js");
+const FilmsAPI = require("./Router/FilmsAPI.js");
+const BookingsAPI = require("./Router/BookingsAPI.js");
+const ScreeningAPI = require("./Router/ScreeningsAPI.js");
 app.use("/api", bodyParser.json(), FilmsAPI, BookingsAPI, ScreeningAPI);
 const server = app.listen(8081, () =>{
     console.log("Connected");
